end of forms/formsets and file/image handling
I had a hard time adapting to work and study, but now I'm okay with it. All this time, I tried to revise my knowledge to not forget anything. But today, I finished learning about forms and formsets with file handling. All that's left is the Pillow library, but I want to study authentication first; it's a big and interesting topic.